Rename repo
- Change all names `Primitive`->`Prime`
- Change Namespace to `\Knowgod\Prime`
