+++ composer.json
@@ -1,6 +1,6 @@
 {
-  "name": "knowgod/primitive-numbers-finder",
-  "description": "primitive numbers finder",
+  "name": "knowgod/prime-numbers-finder",
+  "description": "Prime numbers finder",
   "type": "project",
   "license": "MIT",
   "version": "0.1.0",
@@ -17,7 +17,7 @@
   },
   "autoload": {
     "psr-4": {
-      "Primitive\\": "src/"
+      "Knowgod\\Prime\\": "src/"
     }
   }
 }
